Lots of little fixes Based all level calculations on the decimal level  rather than integer. This allows more accurate representation of your progress. Without this you could up to 99% of each level and the progress would not be reflected until you got the next 1%.
Added in some bad indicators to show watch category each class is.

Sorted out to some degree the formatting, which should work better on
mobile devices, and flow better.

Added lodestone id to the form due to new rate limiting. This is cached.

Caching the character call, to avoid 429s, if you are in the XIV cache
 it should be much quicker, if not, you will likely get nothing back.